Split BidsInputs class two classes
BidsDataset is the new equivalent of BidsInputs, having the same API. In
addition, it subclasses UserDict, and it's substituent components are
now available via dict bracket-syntax.

BidsComponent defines an individual component as defined in
pybids_inputs. It has a similar API as BidsInputs, but for only one
item.

input_lists and input_wildcards are no longer necessary for
construction, but are computed within BidsComponent using
input_zip_lists

BidsComponents are compared to each other by __eq__ via zip_lists, path,
and name.